# Updated cache mapping delegates to remove per-call reflection