Moved activation logic from ToFacet<TSource,TTarget>() into a generic static FacetCache<TSource,TTarget> which compiles and stores a mapping delegate (FromSource factory method or source-based constructor) on first use.

This reduces allocations and removes reflection from the hot path, improving mapping performance

Benchmarked mapping two objects in sequence.

- Before change:

| Method       | Mean     | Error   | StdDev  | Gen0   | Allocated |
|------------- |---------:|--------:|--------:|-------:|----------:|
| MapWithFacet | 354.2 ns | 3.19 ns | 2.98 ns | 0.0701 |    1104 B |
| MapManually  | 229.8 ns | 2.22 ns | 1.97 ns | 0.0570 |     896 B |

- After change:

| Method       | Mean     | Error   | StdDev  | Gen0   | Allocated |
|------------- |---------:|--------:|--------:|-------:|----------:|
| MapWithFacet | 120.2 ns | 1.30 ns | 1.15 ns | 0.0336 |     528 B |
| MapManually  | 229.5 ns | 2.55 ns | 2.13 ns | 0.0570 |     896 B |

The change is backwards-compatible.

The test project ran successfully after this change.